Design Matrices

API

It is often convenient to generate multiple independent sequences, for error estimation (uncertainty quantification). The resulting sequences can be stored in what is often called a design matrix. In this package, this is achieved with the generate_design_matrices(n, d, ::DeterministicSamplingAlgorithm), ::RandomizationMethod, num_mats) function. num_mats is the number of independent realizations. The resulting design matrix is a vector of matrix of length num_mats.

QuasiMonteCarlo.generate_design_matrices — Function

generate_design_matrices(n, d, sample_method::DeterministicSamplingAlgorithm,
    num_mats, T = Float64)
generate_design_matrices(n, d, sample_method::RandomSamplingAlgorithm,
    num_mats, T = Float64)
generate_design_matrices(n, lb, ub, sample_method,
    num_mats = 2)

Create num_mats matrices, each containing a QMC point set, where:

n is the number of points to sample.

d is the dimensionality of the point set in [0, 1)ᵈ,

sample_method is the quasi-Monte Carlo sampling strategy used to create a deterministic point set out.

T is the eltype of the point sets. For some QMC methods (Faure, Sobol) this can be Rational If the bounds lb and ub are specified instead of d, the samples will be transformed into the box [lb, ub].

generate_design_matrices(n, d, sampler, R::NoRand, num_mats, T = Float64)

R = NoRand() produces num_mats matrices, each containing a different deterministic point set in [0, 1)ᵈ. Note that this is an ad hoc way to produce i.i.d. sequences, as it creates a deterministic point in dimension d × num_mats and splits it in num_mats point set of dimension d. This does not have any QMC guarantuees.

Warning

The method generate_design_matrices(n, d, sampler, R::NoRand, num_mats, T = Float64) is an ad hoc way to produce a Design Matrix. Indeed, it creates a deterministic point set in dimension d × num_mats and splits it into num_mats point set of dimension d. The resulting sequences have no QMC guarantees. This seems to have been proposed in Section 5.1 of Saltelli, A. et al. (2010)[1] to do uncertainty quantification.

Example

using QuasiMonteCarlo, Random
Random.seed!(1234)
m = 4
d = 3
b = QuasiMonteCarlo.nextprime(d)
N = b^m # Number of points
pad = m # Can also choose something as `2m` to get "better" randomization

# 5 independent Randomized Faure sequences
QuasiMonteCarlo.generate_design_matrices(N,
    d,
    FaureSample(R = OwenScramble(base = b, pad = m)),
    5)

5-element Vector{Matrix{Float64}}:
 [0.14814814814814814 0.38271604938271603 … 0.6049382716049383 0.8765432098765432; 0.8395061728395062 0.2469135802469136 … 0.30864197530864196 0.43209876543209874; 0.6172839506172839 0.8888888888888888 … 0.024691358024691357 0.3950617283950617]
 [0.5802469135802469 0.7407407407407408 … 0.9259259259259259 0.16049382716049382; 0.7037037037037037 0.25925925925925924 … 0.2222222222222222 0.4444444444444444; 0.5555555555555555 0.6666666666666666 … 0.2469135802469136 0.4197530864197531]
 [0.7901234567901234 0.024691358024691357 … 0.2962962962962963 0.5061728395061729; 0.7530864197530863 0.012345679012345678 … 0.08641975308641975 0.46913580246913583; 0.1728395061728395 0.38271604938271603 … 0.7530864197530863 0.2222222222222222]
 [0.024691358024691357 0.35802469135802467 … 0.6172839506172839 0.8765432098765432; 0.345679012345679 0.8518518518518517 … 0.8271604938271605 0.23456790123456792; 0.19753086419753085 0.45679012345679015 … 0.7283950617283951 0.2962962962962963]
 [0.7283950617283951 0.16049382716049382 … 0.061728395061728385 0.35802469135802467; 0.30864197530864196 0.46913580246913583 … 0.5308641975308642 0.7530864197530863; 0.7654320987654321 0.2962962962962963 … 0.35802469135802467 0.8024691358024691]

1Saltelli, A., Annoni, P., Azzini, I., Campolongo, F., Ratto, M., & Tarantola, S. (2010). Variance based sensitivity analysis of model output. Design and estimator for the total sensitivity index. Computer physics communications, 181(2), 259-270.