Sampler APIs

Sample

Missing docstring.

Missing docstring for sample. Check Documenter's build log for details.

Samplers

Samplers are divided into two subtypes

abstract type SamplingAlgorithm end
abstract type RandomSamplingAlgorithm <: SamplingAlgorithm end
abstract type DeterministicSamplingAlgorithm <: SamplingAlgorithm end

Deterministic Sampling Algorithm

QuasiMonteCarlo.GridSample — Type

GridSample()

A simple rectangular grid lattice.

In more than 2 dimensions, grids have worse discrepancy than simple Monte Carlo. As a result, they should almost never be used for multivariate integration; their use is as a starting point for other algorithms.

QuasiMonteCarlo.SobolSample — Type

SobolSample <: DeterministicSamplingAlgorithm

Samples taken from Sobol's base-2 sequence.

QuasiMonteCarlo.LatticeRuleSample — Type

LatticeRuleSample()

Generate a point set using a lattice rule.

QuasiMonteCarlo.HaltonSample — Type

HaltonSample()

Create a Halton sequence.

QuasiMonteCarlo.GoldenSample — Function

GoldenSample()

Generate a quasirandom Kronecker sequence using powers of the generalized golden ratio.

The harmonious, or generalized golden, ratios are defined as the solutions to the equation: $x^d = x + 1$

Where d is the dimension of the sequence. The Golden sequence is then equivalent to Kronecker([x^-i for i in 1:d]).

WARNING: the generalized golden sequence in more than 2 dimensions is purely experimental. It likely has poor discrepancy in high dimensions, and should not be used without verifying answers against a better-known quasirandom sequence. Try a rank-1 lattice rule instead.

References: Roberts, M. (2018). The Unreasonable Effectiveness of Quasirandom Sequences. Extreme Learning. http://extremelearning.com.au/unreasonable-effectiveness-of-quasirandom-sequences/

QuasiMonteCarlo.KroneckerSample — Type

KroneckerSample(generator::AbstractVector) <: DeterministicSamplingAlgorithm

A Kronecker sequence is a point set generated using a vector and equation: x[i] = i * generator .% 1

Where i runs from 1 through the sample size n. This sequence will be equidistributed (uniform in the infinite limit) so long as the components of generator are linearly independent over the field of rational numbers.

If no generator is specified, a lattice based on the generalized golden ratio is used; see GoldenSample for more information.

Kronecker sequences are not recommended for use in more than 3 dimensions, as theory on them is sparse. LatticeRuleSample will return rank-1 lattice rules, which behave similarly to Kronecker sequences but have better properties.

References: Leobacher, G., & Pillichshammer, F. (2014). Introduction to quasi-Monte Carlo integration and applications. Switzerland: Springer International Publishing. https://link.springer.com/content/pdf/10.1007/978-3-319-03425-6.pdf

Random Sampling Algorithm

Missing docstring.

Missing docstring for UniformSample. Check Documenter's build log for details.

QuasiMonteCarlo.LatinHypercubeSample — Type

LatinHypercubeSample <: RandomSamplingAlgorithm

A Latin Hypercube is a point set with the property that every one-dimensional interval (i/n, i+1/n) contains exactly one point. This is a good way to sample a high-dimensional space, as it is more uniform than a random sample but does not require as many points as a full net.