Sampler APIs

Sample

Missing docstring.

Missing docstring for sample. Check Documenter's build log for details.

Samplers

QuasiMonteCarlo.GridSample — Type

GridSample{T}

The grid is given by lb:dx[i]:ub in the ith direction.

QuasiMonteCarlo.UniformSample — Type

struct UniformSample <: SamplingAlgorithm end

Uniformly distributed random numbers.

QuasiMonteCarlo.SobolSample — Type

struct SobolSample <: SamplingAlgorithm end

Samples using the Sobol sequence using Sobol.jl.

QuasiMonteCarlo.LatinHypercubeSample — Type

struct LatinHypercubeSample <: SamplingAlgorithm end

Samples using a Latin Hypercube using LatinHypercubeSampling.jl

LatinHypercubeSample(threading=false)

Keyword arguments:

threading: whether to use threading. Default is false, i.e. serial.

QuasiMonteCarlo.LatticeRuleSample — Type

struct LatticeRuleSample <: SamplingAlgorithm end

Samples using a randomly-shifted rank-1 lattice rule using LatticeRules.jl

QuasiMonteCarlo.LowDiscrepancySample — Type

struct LowDiscrepancySample{T} <: SamplingAlgorithm

base[i] is the base in the ith direction.

Keyword arguments:

rotation: whether to apply Cranley-Patterson rotation. It can improve Quasi-Monte Carlo integral estimates done with LowDiscrepancy sequences (only Halton, in this case). Defaults to true.

QuasiMonteCarlo.GoldenSample — Type

GoldenSample(origin)

Generate a quasirandom Kronecker sequence using powers of the generalized golden ratio.

The harmonious, or generalized golden, ratios are defined as the solutions to the equation: $x^d = x + 1$

Where d is the dimension of the sequence. The Golden sequence is then equivalent to Kronecker([x^-i for i in 1:d]).

WARNING: the generalized golden sequence in more than 2 dimensions is purely experimental. It likely has poor discrepancy in high dimensions, and should not be used without verifying answers against a better-known quasirandom sequence. Try a rank-1 lattice rule.

References: Roberts, M. (2018). The Unreasonable Effectiveness of Quasirandom Sequences. Extreme Learning. http://extremelearning.com.au/unreasonable-effectiveness-of-quasirandom-sequences/

QuasiMonteCarlo.KroneckerSample — Type

KroneckerSample{V} <: SamplingAlgorithm
KroneckerSample(origin::AbstractVector, generator::AbstractVector)
KroneckerSample(d::Integer, args...)

A Kronecker sequence is a point set generated by a pair of vectors and the equation: x[i] = (origin + i * generator) .% 1

Where i runs from 1 through the sample size n. This sequence will be equidistributed (uniform in the infinite limit) so long as the components of generator are linearly independent over the field of rational numbers.

If origin is not specified, it is assumed to be equal to 0. If only a dimension is specified, a parametrization based on the generalized golden ratio is used, but this is known to give poor samples for dimensions higher than 3–see

Kronecker sequences are not recommended for use in more than 3 dimensions, as theory on them is sparse. LatticeRuleSample will return rank-1 lattice rules, which behave similarly to Kronecker sequences but have better-understood properties.

References: Leobacher, G., & Pillichshammer, F. (2014). Introduction to quasi-Monte Carlo integration and applications. Switzerland: Springer International Publishing. https://link.springer.com/content/pdf/10.1007/978-3-319-03425-6.pdf

QuasiMonteCarlo.SectionSample — Type

struct SectionSample{T} <: SamplingAlgorithm

SectionSample(x0, sampler) where sampler is any sampler above and x0 is a vector of either NaN for a free dimension or some scalar for a constrained dimension.